refactor for separating several hooks
