+++ src/components/dapps/ModalCreateDapps.vue
@@ -319,18 +319,14 @@ import IconSolidSelector from '@/components/icons/IconSolidSelector.vue';
 // import DeploymentAccountSelectOption from '@/components/dapps/DeploymentAccountSelectOption.vue';
 import type { SubmittableExtrinsic } from '@polkadot/api/types';
 import type { CodeSubmittableResult } from '@polkadot/api-contract/promise/types';
-import type {
-  PartialQueueTxExtrinsic,
-  QueueTxExtrinsic,
-  QueueTx,
-  QueueTxRpc,
-} from '@/types/Status';
+import type { QueueTx } from '@/types/Status';
+import type { ApiPromise } from '@polkadot/api';
 import BN from 'bn.js';
 import * as plasmUtils from '@/helper';
 import ModalSelectAccountOption from '@/components/balance/ModalSelectAccountOption.vue';
 // import CategoryMultiSelect from '@/components/dapps/CategoryMultiSelect.vue';
 import InputFile from '@/components/dapps/InputFile.vue';
-import { compactAddLength, isWasm, stringify } from '@polkadot/util';
+import { stringify } from '@polkadot/util';
 import { SubmittableResult } from '@polkadot/api';
 import { ActionTypes } from '@/store/action-types';
 import { MutationTypes } from '@/store/mutation-types';
@@ -339,13 +335,13 @@ import { useApi } from '@/hooks';
 import useFile, { FileState } from '@/hooks/useFile';
 import useAbi from '@/hooks/useAbi';
 import useSendTx from '@/hooks/signer/useSendTx';
+import useWasm from '@/hooks/useWasm';
+import usePendingTx from '@/hooks/signer/usePendingTx';
 import { CodePromise, Abi } from '@polkadot/api-contract';
 import { AnyJson } from '@polkadot/types/types';
-import jsonrpc from '@polkadot/types/interfaces/jsonrpc';
 import { useStore } from 'vuex';
 import { AddressProxy } from '@/types/Signer';
 import { bnToBn } from '@polkadot/util';
-import type { ApiPromise } from '@polkadot/api';
 import { getParamValues } from '@/helper/params';
 
 interface FormData {
@@ -441,35 +437,16 @@ export default defineComponent({
         file?.data.subarray(0, 4).toString() === '0,97,115,109',
     });
 
-    const wasm = ref();
-    const isWasmValid = ref(false);
     const extensionFile = ['.contract', '.json'];
 
     const onDropFile = (fileState: FileState) => {
       setWasmFile(fileState);
     };
 
-    watch(abi, () => {
-      console.log('abc', abi.value);
-      console.log('wasm', abi.value?.project.source.wasm);
-      console.log('constructor', abi.value?.constructors);
-      if (abi.value && isWasm(abi.value.project.source.wasm)) {
-        wasm.value = abi.value.project.source.wasm;
-        isWasmValid.value = true;
-
-        return;
-      }
-
-      if (wasmFromFile.value && isWasmFromFileValid) {
-        wasm.value = compactAddLength(wasmFromFile.value.data);
-        isWasmValid.value = true;
-
-        return;
-      }
-
-      wasm.value = null;
-      isWasmValid.value = false;
-    });
+    const { wasm, isWasmValid } = useWasm(
+      wasmFromFile.value,
+      isWasmFromFileValid
+    );
 
     const { onSend } = useSendTx();
 
@@ -477,7 +454,8 @@ export default defineComponent({
       if (
         formData.projectName === '' ||
         toAddress.value === '' ||
-        wasm.value === null
+        wasm.value === null ||
+        !isWasmValid
       ) {
         store.dispatch(ActionTypes.SHOW_ALERT_MSG, {
           msg: `Please check fields again`,
@@ -542,19 +520,6 @@ export default defineComponent({
       console.log('uploadTx', uploadTx);
 
       //send from txButton
-      const SUBMIT_RPC = jsonrpc.author.submitAndWatchExtrinsic;
-      const propsExtrinsic = uploadTx;
-
-      let extrinsics: SubmittableExtrinsic<'promise'>[] | undefined;
-
-      if (propsExtrinsic) {
-        extrinsics = Array.isArray(propsExtrinsic)
-          ? propsExtrinsic
-          : [propsExtrinsic];
-      }
-
-      const accountId = props.address;
-
       const _onFailed = (result: SubmittableResult | null) => {
         console.error('_onFailed', result);
         store.commit(MutationTypes.SET_LOADING, false);
@@ -608,42 +573,14 @@ export default defineComponent({
       };
       const _onUpdate = () => {};
 
-      let txqueue: QueueTx[] = [];
-      let nextId = 0;
-
-      const addToTxQueue = (
-        value: QueueTxExtrinsic | QueueTxRpc | QueueTx
-      ): void => {
-        const id = ++nextId;
-        const removeItem = () => {
-          ////
-        };
-
-        txqueue = [
-          {
-            ...value,
-            id,
-            removeItem,
-            rpc: (value as QueueTxRpc).rpc || SUBMIT_RPC,
-            status: 'queued',
-          },
-        ];
-      };
-
-      const queueExtrinsic = (value: PartialQueueTxExtrinsic) =>
-        addToTxQueue({ ...value });
-
-      extrinsics?.forEach((extrinsic): void => {
-        queueExtrinsic({
-          accountId: accountId && accountId.toString(),
-          extrinsic,
-          isUnsigned: false,
-          txFailedCb: _onFailed,
-          txStartCb: _onStart,
-          txSuccessCb: _onSuccess,
-          txUpdateCb: _onUpdate,
-        });
-      });
+      const { txqueue } = usePendingTx(
+        uploadTx,
+        props.address,
+        _onStart,
+        _onFailed,
+        _onSuccess,
+        _onUpdate
+      );
 
       console.log('txQueue', txqueue[0]);
